Update relative paths in start.sh script for MCE installs
This installation script must have been shuffled around from the root of
the repository since it contained relative paths to files and
directories available from the repository root. Following the directions
in the multiclusterengine/ directory results in a failed deploy because
the script assumes those relative paths.

This commit updates the script so that it installs cleanly by updating
the relative paths.

Closes #242